requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pymongo==4.10.1
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.12